+++ packages/atomic/src/components.d.ts
@@ -712,6 +712,10 @@ export namespace Components {
          */
         "initialize": (options: InitializationOptions) => Promise<void>;
         /**
+          * Initializes the connection with an already preconfigured headless search engine, as opposed to the `initialize` method which will internally create a new search engine instance.
+         */
+        "initializeWithSearchEngine": (engine: SearchEngine) => Promise<void>;
+        /**
           * The search interface language.
          */
         "language": string;