+++ packages/headless/src/features/search-and-folding/search-and-folding-request.ts
@@ -1,7 +1,8 @@
 import {SearchAppState} from '../../state/search-app-state';
-import {getVisitorID} from '../../api/analytics/analytics';
+import {getVisitorID, historyStore} from '../../api/analytics/analytics';
 import {ConfigurationSection} from '../../state/state-sections';
 import {fromAnalyticsStateToAnalyticsParams} from '../configuration/configuration-state';
+import {isNullOrUndefined} from '@coveo/bueno';
 
 type StateNeededByExecuteSearchAndFolding = ConfigurationSection &
   Partial<SearchAppState>;
@@ -20,6 +21,7 @@ export const buildSearchAndFoldingLoadCollectionRequest = async (
     timezone: state.configuration.search.timezone,
     ...(state.configuration.analytics.enabled && {
       visitorId: await getVisitorID(),
+      actionsHistory: historyStore.getHistory(),
     }),
     ...(state.advancedSearchQueries?.aq && {
       aq: state.advancedSearchQueries.aq,
@@ -57,5 +59,9 @@ export const buildSearchAndFoldingLoadCollectionRequest = async (
       (await fromAnalyticsStateToAnalyticsParams(
         state.configuration.analytics
       ))),
+    ...(state.excerptLength &&
+      !isNullOrUndefined(state.excerptLength.length) && {
+        excerptLength: state.excerptLength.length,
+      }),
   };
 };