+++ packages/samples/vuejs/src/main.ts
@@ -1,4 +1,10 @@
-import {createApp} from 'vue';
+import {createApp, defineCustomElement} from 'vue';
 import App from './App.vue';
+import ResultTextField from './components/ResultTextField.vue';
 
-createApp(App).mount('#app');
+const resultTextField = defineCustomElement(ResultTextField);
+customElements.define('result-text-field', resultTextField);
+
+const app = createApp(App);
+
+app.mount('#app');